code/deployment/api/requirements.txt
pillow
numpy
opencv-python-headless
onnxruntime
fastapi
pydantic
gunicorn
uvicorn[standard]
python-multipart
gunicorn
